requirements.txt
opencv-python-headless
Pillow
numpy
imageio
